@@ -3,7 +3,7 @@
 
   <PropertyGroup>
     <RootNamespace>ClangSharp</RootNamespace>
-    <TargetFrameworks>net8.0</TargetFrameworks>
+    <TargetFrameworks>net10.0</TargetFrameworks>
   </PropertyGroup>
 
   <PropertyGroup>
