@@ -3,13 +3,15 @@
 
   <PropertyGroup>
     <RootNamespace>ClangSharp.UnitTests</RootNamespace>
-    <TargetFrameworks>net8.0</TargetFrameworks>
+    <TargetFrameworks>net10.0</TargetFrameworks>
   </PropertyGroup>
 
   <PropertyGroup>
+    <!-- CA1515: Consider making public types internal -->
     <!-- CA1707: Identifiers should not contain underscores -->
     <!-- CA1711: Identifiers should not have incorrect suffix -->
-    <NoWarn>$(NoWarn);CA1707;CA1711</NoWarn>
+    <!-- IDE0130: Namespace does not match folder structure -->
+    <NoWarn>$(NoWarn);CA1515;CA1707;CA1711;IDE0130</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
