@@ -2,12 +2,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net8.0</TargetFrameworks>
+    <TargetFrameworks>net10.0</TargetFrameworks>
   </PropertyGroup>
 
   <PropertyGroup>
+    <!-- CA1515: Consider making public types internal -->
     <!-- CA1707: Identifiers should not contain underscores -->
-    <NoWarn>$(NoWarn);CA1707</NoWarn>
+    <!-- IDE0130: Namespace does not match folder structure -->
+    <NoWarn>$(NoWarn);CA1515;CA1707;IDE0130</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
